// packages/chapterjs/src/discord/types/message.ts
export interface RawMessage {
  /** id of the message */
  id: Snowflake;
  /** id of the channel the message was sent in */
  channel_id: Snowflake;
  /** the author of this message (not guaranteed to be a valid user, see below) */
  author: RawUser;
  /** contents of the message */
  content: string;
  /** when this message was sent */
  timestamp: ISO8601Timestamp;
  /** when this message was edited (or null if never) */
  edited_timestamp: ISO8601Timestamp | null;
  /** whether this was a TTS message */
  tts: boolean;
  /** whether this message mentions everyone */
  mention_everyone: boolean;
  /** users specifically mentioned in the message */
  mentions: RawUser[];
  /** roles specifically mentioned in this message */
  mention_roles: Snowflake[];
  /** channels specifically mentioned in this message */
  mention_channels?: RawChannelMention[];
  /** any attached files that are not referenced in embeds or components */
  attachments: RawAttachment[];
  /** any embedded content */
  embeds: RawEmbed[];
  /** reactions to the message */
  reactions?: RawReaction[];
  /** used for validating a message was sent */
  nonce?: number | string;
  /** whether this message is pinned */
  pinned: boolean;
  /** if the message is generated by a webhook, this is the webhook's id */
  webhook_id?: Snowflake;
  /** type of message */
  type: MessageType;
  /** sent with Rich Presence-related chat embeds */
  activity?: RawMessageActivity;
  /** sent with Rich Presence-related chat embeds */
  application?: Partial<RawApplication>;
  /** if the message is an Interaction or application-owned webhook, this is the id of the application */
  application_id?: Snowflake;
  /** message flags combined as a bitfield */
  flags?: number;
  /** data showing the source of a crosspost, channel follow add, pin, or reply message */
  message_reference?: RawMessageReference;
  /** the message associated with the `message_reference`. This is a minimal subset of fields in a message (e.g. `author` is excluded.) */
  message_snapshots?: RawMessageSnapshot[];
  /** the message associated with the message_reference */
  referenced_message?: RawMessage | null;
  /** Sent if the message is sent as a result of an interaction */
  interaction_metadata?: RawMessageInteractionMetadata;
  /** **Deprecated in favor of `interaction_metadata`**; sent if the message is a response to an interaction */
  interaction?: RawMessageInteraction;
  /** the thread that was started from this message, includes thread member object */
  thread?: RawChannel;
  /** sent if the message contains components like buttons, action rows, or other interactive components */
  components?: RawComponent[];
  /** sent if the message contains stickers */
  sticker_items?: RawStickerItem[];
  /** **Deprecated** the stickers sent with the message */
  stickers?: RawSticker[];
  /** A generally increasing integer (there may be gaps or duplicates) that represents the approximate position of the message in a thread, it can be used to estimate the relative position of the message in a thread in company with `total_message_sent` on parent thread */
  position?: number;
  /** data of the role subscription purchase or renewal that prompted this ROLE_SUBSCRIPTION_PURCHASE message */
  role_subscription_data?: RawRoleSubscriptionData;
  /** data for users, members, channels, and roles referenced in this message */
  resolved?: RawResolvedData;
  /** A poll! */
  poll?: RawPoll;
  /** the call associated with the message */
  call?: RawMessageCall;
  /** the custom client-side theme shared via the message */
  shared_client_theme?: RawSharedClientTheme;
}

/**
 * Message Types
 * @see https://docs.discord.com/developers/resources/message#message-object-message-types
 */
export const MessageType = {
  Default: 0,
  RecipientAdd: 1,
  RecipientRemove: 2,
  Call: 3,
  ChannelNameChange: 4,
  ChannelIconChange: 5,
  ChannelPinnedMessage: 6,
  UserJoin: 7,
  GuildBoost: 8,
  GuildBoostTier1: 9,
  GuildBoostTier2: 10,
  GuildBoostTier3: 11,
  ChannelFollowAdd: 12,
  GuildDiscoveryDisqualified: 14,
  GuildDiscoveryRequalified: 15,
  GuildDiscoveryGracePeriodInitialWarning: 16,
  GuildDiscoveryGracePeriodFinalWarning: 17,
  ThreadCreated: 18,
  Reply: 19,
  ChatInputCommand: 20,
  ThreadStarterMessage: 21,
  GuildInviteReminder: 22,
  ContextMenuCommand: 23,
  AutoModerationAction: 24,
  RoleSubscriptionPurchase: 25,
  InteractionPremiumUpsell: 26,
  StageStart: 27,
  StageEnd: 28,
  StageSpeaker: 29,
  StageTopic: 31,
  GuildApplicationPremiumSubscription: 32,
  GuildIncidentAlertModeEnabled: 36,
  GuildIncidentAlertModeDisabled: 37,
  GuildIncidentReportRaid: 38,
  GuildIncidentReportFalseAlarm: 39,
  PurchaseNotification: 44,
  PollResult: 46,
} as const;
export type MessageType = (typeof MessageType)[keyof typeof MessageType];

/**
 * Message Activity Structure
 * @see https://docs.discord.com/developers/resources/message#message-object-message-activity-structure
 */
export interface RawMessageActivity {
  /** type of message activity */
  type: MessageActivityType;
  /** party_id from a Rich Presence event */
  party_id?: string;
}
// ...
export const MessageFlags = {
  /** this message has been published to subscribed channels (via Channel Following) */
  Crossposted: 1 << 0,
  /** this message originated from a message in another channel (via Channel Following) */
  IsCrosspost: 1 << 1,
  /** do not include any embeds when serializing this message */
  SuppressEmbeds: 1 << 2,
  /** the source message for this crosspost has been deleted (via Channel Following) */
  SourceMessageDeleted: 1 << 3,
  /** this message came from the urgent message system */
  Urgent: 1 << 4,
  /** this message has an associated thread, with the same id as the message */
  HasThread: 1 << 5,
  /** this message is only visible to the user who invoked the Interaction */
  Ephemeral: 1 << 6,
  /** this message is an Interaction Response and the bot is "thinking" */
  Loading: 1 << 7,
  /** this message failed to mention some roles and add their members to the thread */
  FailedToMentionSomeRolesInThread: 1 << 8,
  /** this message will not trigger push and desktop notifications */
  SuppressNotifications: 1 << 12,
  /** this message is a voice message */
  IsVoiceMessage: 1 << 13,
  /** this message has a snapshot (via Message Forwarding) */
  HasSnapshot: 1 << 14,
  /** allows you to create fully component-driven messages */
  IsComponentsV2: 1 << 15,
} as const;
export type MessageFlags = (typeof MessageFlags)[keyof typeof MessageFlags];
// ...
export interface RawMessageCall {
  /** array of user object ids that participated in the call */
  participants: Snowflake[];
  /** time when call ended */
  ended_timestamp?: ISO8601Timestamp | null;
}

/**
 * Message Reference Structure
 * @see https://docs.discord.com/developers/resources/message#message-reference-structure
 */
export interface RawMessageReference {
  /** type of reference. */
  type?: MessageReferenceType;
  /** id of the originating message */
  message_id?: Snowflake;
  /** id of the originating message's channel */
  channel_id?: Snowflake;
  /** id of the originating message's guild */
  guild_id?: Snowflake;
  /** when sending, whether to error if the referenced message doesn't exist instead of sending as a normal (non-reply) message, default true */
  fail_if_not_exists?: boolean;
}
// ...
export interface RawMessageSnapshot {
  /** minimal subset of fields in the forwarded message */
  message: Partial<RawMessage>;
}

/**
 * Reaction Structure
 * @see https://docs.discord.com/developers/resources/message#reaction-object-reaction-structure
 */
export interface RawReaction {
  /** Total number of times this emoji has been used to react (including super reacts) */
  count: number;
  /** Reaction count details object */
  count_details: RawReactionCountDetails;
  /** Whether the current user reacted using this emoji */
  me: boolean;
  /** Whether the current user super-reacted using this emoji */
  me_burst: boolean;
  /** emoji information */
  emoji: Partial<RawEmoji>;
  /** HEX colors used for super reaction */
  burst_colors: string[];
}

/**
 * Reaction Count Details Structure
 * @see https://docs.discord.com/developers/resources/message#reaction-count-details-object-reaction-count-details-structure
 */
export interface RawReactionCountDetails {
  /** Count of super reactions */
  burst: number;
  /** Count of normal reactions */
  normal: number;
}

/**
 * Embed Structure
 * @see https://docs.discord.com/developers/resources/message#embed-object-embed-structure
 */
export interface RawEmbed {
  /** title of embed */
  title?: string;
  /** type of embed (always "rich" for webhook embeds) */
  type?: EmbedType;
  /** description of embed */
  description?: string;
  /** url of embed */
  url?: string;
  /** timestamp of embed content */
  timestamp?: ISO8601Timestamp;
  /** color code of the embed */
  color?: number;
  /** footer information */
  footer?: RawEmbedFooter;
  /** image information */
  image?: RawEmbedImage;
  /** thumbnail information */
  thumbnail?: RawEmbedImage;
  /** video information */
  video?: RawEmbedVideo;
  /** provider information */
  provider?: RawEmbedProvider;
  /** author information */
  author?: RawEmbedAuthor;
  /** fields information, max of 25 */
  fields?: RawEmbedField[];
  /** embed flags combined as a bitfield */
  flags?: number;
}
// ...
export interface RawEmbedProvider {
  /** name of provider */
  name?: string;
  /** url of provider */
  url?: string;
}

/**
 * Embed Author Structure
 * @see https://docs.discord.com/developers/resources/message#embed-object-embed-author-structure
 */
export interface RawEmbedAuthor {
  /** name of author */
  name: string;
  /** url of author (only supports http(s)) */
  url?: string;
  /** url of author icon (only supports http(s) and attachments) */
  icon_url?: string;
  /** a proxied url of author icon */
  proxy_icon_url?: string;
}

/**
 * Embed Footer Structure
 * @see https://docs.discord.com/developers/resources/message#embed-object-embed-footer-structure
 */
export interface RawEmbedFooter {
  /** footer text */
  text: string;
  /** url of footer icon (only supports http(s) and attachments) */
  icon_url?: string;
  /** a proxied url of footer icon */
  proxy_icon_url?: string;
}
// ...
export interface RawAttachment {
  /** attachment id */
  id: Snowflake;
  /** name of file attached */
  filename: string;
  /** the title of the file */
  title?: string;
  /** description (alt text) for the file (max 1024 characters) */
  description?: string;
  /** the attachment's media type */
  content_type?: string;
  /** size of file in bytes */
  size: number;
  /** source url of file */
  url: string;
  /** a proxied url of file */
  proxy_url: string;
  /** height of file (if image or video) */
  height?: number | null;
  /** width of file (if image or video) */
  width?: number | null;
  /** thumbhash placeholder (if image or video) */
  placeholder?: string;
  /** version of the placeholder (if image or video) */
  placeholder_version?: number;
  /** whether this attachment is ephemeral */
  ephemeral?: boolean;
  /** the duration of the audio or video file */
  duration_secs?: number;
  /** base64 encoded bytearray representing a sampled waveform (currently for voice messages) */
  waveform?: string;
  /** attachment flags combined as a bitfield */
  flags?: number;
  /** for Clips, array of users who were in the stream */
  clip_participants?: RawUser[];
  /** for Clips, when the clip was created */
  clip_created_at?: ISO8601Timestamp;
  /** for Clips, the application in the stream, if recognized */
  application?: Partial<RawApplication> | null;
}
// ...
export interface RawChannelMention {
  /** id of the channel */
  id: Snowflake;
  /** id of the guild containing the channel */
  guild_id: Snowflake;
  /** the type of channel */
  type: ChannelType;
  /** the name of the channel */
  name: string;
}
// ...
export interface RawRoleSubscriptionData {
  /** the id of the sku and listing that the user is subscribed to */
  role_subscription_listing_id: Snowflake;
  /** the name of the tier that the user is subscribed to */
  tier_name: string;
  /** the cumulative number of months that the user has been subscribed for */
  total_months_subscribed: number;
  /** whether this notification is for a renewal rather than a new purchase */
  is_renewal: boolean;
}
// ...
export interface RawSharedClientTheme {
  /** the hexadecimal-encoded colors of the theme (max of 5) */
  colors: string[];
  /** the direction of the theme's colors (max of 360) */
  gradient_angle: number;
  /** the intensity of the theme's colors (max of 100) */
  base_mix: number;
  /** the mode of the theme */
  base_theme?: SharedClientThemeBaseType | null;
}
// ...
export type EmbedType =
  'rich' | 'image' | 'video' | 'gifv' | 'article' | 'link' | 'poll_result';

/**
 * Message Interaction Metadata: its shape depends on the interaction type.
 * @see https://docs.discord.com/developers/resources/message#message-interaction-metadata-object
 */
export type RawMessageInteractionMetadata =
  | RawApplicationCommandInteractionMetadata
  | RawMessageComponentInteractionMetadata
  | RawModalSubmitInteractionMetadata;